Produce and send to console a tibble of the data request codes and associated datasets.

`show_datacodes()` produces a tibble of the arguments used with the `get_ausdata()` function and the associated datasets.

show_datacodes(limit = 4)

Arguments

limit
A numeric value used to determine the number of values returned. Default is set to four (4) values, which returns all tibble values.

Value

A console printout of a tibble with a designated number of values set by limit.

Details

Items under *Request Codes* can be used to request and download the associated Australian Politicians dataset.

The specifics of these are:

- `"all"` australian_politicians-all.csv. - `"allbyparty"` australian_politicians-all-by_party.csv. - `"mps"` australian_politicians-mps-by_division.csv. - `"senators"` australian_politicians-senators-by_state.csv.

See also

Examples

if (FALSE) {
# Print out dataset request codes.
show_datacodes()

# Request "All" Australian Politicians dataset.
get_auspol("all")
}